Fix up some typing and display issues

+++ Infrastructure/Persistence/Repositories/DeviceStatusRepository.cs
@@ -68,12 +68,12 @@ namespace RightpointLabs.ConferenceRoom.Infrastructure.Persistence.Repositories
             public string OrganizationId { get; set; }
             public string DeviceId { get; set; }
             public DateTime StatusTimestamp { get; set; }
-            public decimal? Temperature1 { get; set; }
-            public decimal? Temperature2 { get; set; }
-            public decimal? Temperature3 { get; set; }
-            public decimal? Voltage1 { get; set; }
-            public decimal? Voltage2 { get; set; }
-            public decimal? Voltage3 { get; set; }
+            public double? Temperature1 { get; set; }
+            public double? Temperature2 { get; set; }
+            public double? Temperature3 { get; set; }
+            public double? Voltage1 { get; set; }
+            public double? Voltage2 { get; set; }
+            public double? Voltage3 { get; set; }
 
             public DeviceStatusEntity()
             {
